requirements.txt
pyosc
tensorflow
magenta
